backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
huggingface-hub==0.17.3
transformers==4.35.2
torch==2.1.1
beautifulsoup4==4.12.2
nltk==3.8.1
pydantic==2.5.1
motor==3.3.2
aiohttp==3.9.1
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
textstat==0.7.3
aiofiles==23.2.1
python-dateutil==2.8.2